+++ packages/react/src/atomic/Divider/src/divider.types.ts
@@ -11,7 +11,7 @@ export interface DividerProperties {
    * Visual appearance color.
    * @default neutral
    */
-  appearance?: "neutral";
+  appearance?: "neutral" | "primary";
 
   /**
    * Thickness level based on shape.border.width tokens.